Extract linear_extrude-related functions into a separate file
